+++ webgui.h
@@ -17,16 +17,16 @@ string make_in_link(in*);
 string make_out_link(out*);
 
 // in*, tmin, tmax, x-pixels, y-pixels
-string plotLine(in*, unsigned long, unsigned long, unsigned int, unsigned int);
+string plotLine(in*, double, double, unsigned int, unsigned int);
 
 // list<in*>, tmin, tmax, x-pixels, y-pixels, title
-string plotLines(list<in*>, unsigned long, unsigned long, unsigned int, unsigned int, string);
+string plotLines(list<in*>, double, double, unsigned int, unsigned int, string);
 // convenience functions, 2, 3 and 4 ins instead of list
-string plotLines(in*, unsigned long, unsigned long, unsigned int, unsigned int, string);
-string plotLines(in*, in*, unsigned long, unsigned long, unsigned int, unsigned int, string);
-string plotLines(in*, in*, in*, unsigned long, unsigned long, unsigned int, unsigned int, string);
-string plotLines(in*, in*, in*, in*, unsigned long, unsigned long, unsigned int, unsigned int, string);
-string plotLines(in*, in*, in*, in*, in*, unsigned long, unsigned long, unsigned int, unsigned int, string);
+string plotLines(in*, double, double, unsigned int, unsigned int, string);
+string plotLines(in*, in*, double, double, unsigned int, unsigned int, string);
+string plotLines(in*, in*, in*, double, double, unsigned int, unsigned int, string);
+string plotLines(in*, in*, in*, in*, double, double, unsigned int, unsigned int, string);
+string plotLines(in*, in*, in*, in*, in*, double, double, unsigned int, unsigned int, string);
 
 // web server management
 void webGuiStart(string port = "8090");
